fix(suppliers): UpdateSupplierServices etait un stub + tri Rating incompatible SQLite
- UpdateSupplierServicesCommandHandler ne remplacait jamais les services
  (commentaire "table doesn't exist yet" perime depuis la migration
  20260125230000_AddSupplierServicesAndOptimizations) : suppression des
  services existants + insertion de la nouvelle liste validee/normalisee,
  comme dans CreateSupplierCommandHandler.
- GetGaragesQueryHandler : ORDER BY sur decimal non supporte par SQLite
  (harnais de tests) ; cast en double, ordre identique sous PostgreSQL.

Tests: 394/394 verts (avant: 392/394).

+++ services/src/GisAPI.Application/Features/Suppliers/Queries/GetGaragesQueryHandler.cs
@@ -43,8 +43,9 @@ public class GetGaragesQueryHandler : IRequestHandler<GetGaragesQuery, Paginated
             query = query.Where(s => s.IsActive == request.IsActive.Value);
         }
 
-        // Order by rating desc, then name
-        query = query.OrderByDescending(s => s.Rating).ThenBy(s => s.Name);
+        // Order by rating desc, then name. Cast en double : SQLite (tests) ne sait pas
+        // trier un decimal ; l'ordre est identique sous PostgreSQL.
+        query = query.OrderByDescending(s => (double)s.Rating).ThenBy(s => s.Name);
 
         var totalCount = await query.CountAsync(cancellationToken);
